fix: move picking mocks
