+++ apps/storybook/.storybook/app/build.ts
@@ -9,7 +9,7 @@ import { formFeature } from '@spryker-oryx/form';
 import { I18nFeature } from '@spryker-oryx/i18n';
 import { mockOfflineFeature } from '@spryker-oryx/offline/mocks';
 import { mockOrderFeature } from '@spryker-oryx/order/mocks';
-import { mockPickingFeature } from '@spryker-oryx/picking/src/mocks';
+import { mockPickingFeature } from '@spryker-oryx/picking/mocks';
 import { mockProductFeature } from '@spryker-oryx/product/mocks';
 import { mockSearchFeature } from '@spryker-oryx/search/mocks';
 import { mockSiteFeature } from '@spryker-oryx/site/mocks';
