+++ libs/domain/picking/customer-note/customer-note.component.spec.ts
@@ -2,7 +2,7 @@ import { fixture } from '@open-wc/testing-helpers';
 import { useComponent } from '@spryker-oryx/core/utilities';
 import { createInjector, destroyInjector } from '@spryker-oryx/di';
 import { PickingListError, PickingListService } from '@spryker-oryx/picking';
-import { mockPickingListData } from '@spryker-oryx/picking/src/mocks';
+import { mockPickingListData } from '@spryker-oryx/picking/mocks';
 import { RouterService } from '@spryker-oryx/router';
 import { modalComponent } from '@spryker-oryx/ui/modal';
 import { html } from 'lit';
