+++ libs/domain/picking/mocks/src/mock-picking-list.service.ts
