+++ libs/domain/picking/offline/services/adapter/picking-list-online-default.adapter.spec.ts
@@ -7,7 +7,7 @@ import {
   PickingListDefaultAdapter,
   PickingListStatus,
 } from '@spryker-oryx/picking';
-import { mockPickingListData } from '@spryker-oryx/picking/src/mocks';
+import { mockPickingListData } from '@spryker-oryx/picking/mocks';
 import { nextTick } from '@spryker-oryx/utilities';
 import { Table } from 'dexie';
 import { of } from 'rxjs';
