+++ libs/domain/picking/picking-product-card/picking-product-card.component.spec.ts
@@ -2,7 +2,7 @@ import { fixture } from '@open-wc/testing-helpers';
 import { QuantityInputComponent } from '@spryker-oryx/cart/quantity-input';
 import { useComponent } from '@spryker-oryx/core/utilities';
 import { ItemsFilters, PickingListItem } from '@spryker-oryx/picking';
-import { mockPickingListData } from '@spryker-oryx/picking/src/mocks';
+import { mockPickingListData } from '@spryker-oryx/picking/mocks';
 import { ImageComponent } from '@spryker-oryx/ui/image';
 import { html } from 'lit';
 import { PickingProductCardComponent } from './picking-product-card.component';
